etl/requirements.txt
python-dotenv==0.13.0
psycopg2-binary==2.9
elasticsearch[async]==8.13.2